Work on the implementation of adding a director to films and additional refactoring

+++ src/main/java/ru/yandex/practicum/filmorate/controller/film/MpaController.java
@@ -22,7 +22,7 @@ public class MpaController {
     }
 
     @GetMapping("/{id}")
-    public MpaRating getMpaById(@PathVariable int id) {
+    public MpaRating getMpaById(@PathVariable long id) {
         return mpaService.getMpaById(id);
     }
 }
